+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - Flash viewer - bitmap stroke style, strokes scaling, cropped strokes
 - Flash viewer - filters zooming
+- Flash viewer - miter strokes
+- SVG export - miter strokes as miter-clip style
 
 ## [14.2.1] - 2021-03-13
 ### Added